+++ main.js
@@ -6,11 +6,11 @@ import { races, racialTrait, randomMinorTrait, biomes, planetTraits } from './ra
 import { defineResources, resource_values, spatialReasoning, craftCost, plasmidBonus, tradeRatio, craftingRatio, crateValue, containerValue, tradeSellPrice, tradeBuyPrice, atomic_mass } from './resources.js';
 import { defineJobs, job_desc, loadFoundry } from './jobs.js';
 import { defineGovernment, defineGarrison, garrisonSize, armyRating, buildQueue } from './civics.js';
-import { renderFortress, bloodwar } from './portal.js';
 import { actions, updateDesc, challengeGeneHeader, challengeActionHeader, checkCityRequirements, checkTechRequirements, checkOldTech, addAction, storageMultipler, checkAffordable, drawCity, drawTech, gainTech, removeAction, evoProgress, housingLabel, oldTech, f_rate, setPlanet, resQueue } from './actions.js';
 import { space, deepSpace, fuel_adjust, int_fuel_adjust, zigguratBonus, setUniverse, universe_types } from './space.js';
-import { events } from './events.js';
+import { renderFortress, bloodwar } from './portal.js';
 import { arpa } from './arpa.js';
+import { events } from './events.js';
 
 var intervals = {};
 
@@ -883,7 +883,8 @@ function fastLoop(){
                 active = global.space.swarm_control.s_max;
             }
             global.space.swarm_control.support = active;
-            let output = powerModifier(active);
+            let solar = global.tech.swarm >= 4 ? (global.tech.swarm >= 5 ? 0.65 : 0.5) : 0.35;
+            let output = powerModifier(active * solar);
             max_power -= output;
             power_grid += output;
             power_generated[loc('space_sun_swarm_satellite_title')] = output;
@@ -1684,6 +1685,10 @@ function fastLoop(){
             know_bd[loc('job_scientist')] = scientist_base + 'v';
             know_bd[loc('hunger')] = ((hunger - 1) * 100) + '%';
             know_bd[loc('tech_sundial')] = sundial_base + 'v';
+            if (global.race['inspired']){
+                know_bd[loc('event_inspiration_bd')] = '100%';
+                delta *= 2;
+            }
             if (global.city['library']){
                 know_bd[loc('city_library')] = ((library_mult - 1) * 100) + '%';
             }
@@ -2194,7 +2199,7 @@ function fastLoop(){
             let delta = graphene_production * 0.6 * zigguratBonus() * hunger * global_multiplier;
 
             let graphene_bd = {};
-            graphene_bd[loc('interstellar_g_factory_bd')] = graphene_production + 'v';
+            graphene_bd[loc('interstellar_g_factory_bd')] = (graphene_production * zigguratBonus()) + 'v';
             graphene_bd[loc('hunger')] = ((hunger - 1) * 100) + '%';
             breakdown.p['Graphene'] = graphene_bd;
             modRes('Graphene', delta * time_multiplier);
@@ -3663,6 +3668,13 @@ function midLoop(){
         if (global.space['gps'] && global.space.gps.count >= 4){
             global.city.market.mtrade += global.space.gps.count * 2;
         }
+
+        if (global.race['inspired']){
+            global.race['inspired']--;
+            if (global.race['inspired'] <= 0){
+                delete global.race['inspired'];
+            }
+        }
         
         let pop_loss = global.resource[global.race.species].amount - caps[global.race.species];
         if (pop_loss > 0){
@@ -3841,7 +3853,7 @@ function midLoop(){
         }
 
         if (global.space['swarm_control']){
-            global.space.swarm_control.s_max = global.space.swarm_control.count * (global.tech['swarm'] && global.tech['swarm'] >= 2 ? 6 : 4);
+            global.space.swarm_control.s_max = global.space.swarm_control.count * (global.tech['swarm'] && global.tech['swarm'] >= 2 ? 18 : 10);
         }
 
         if (global.arpa['sequence'] && global.arpa.sequence.on && gene_sequence){
@@ -3965,6 +3977,7 @@ function midLoop(){
             let spent = { t: 0, r: {}};
             for (let i=0; i<global.queue.queue.length; i++){
                 let struct = global.queue.queue[i];
+                time = global.settings.qAny ? 0 : time;
 
                 let t_action = false;
                 if (deepScan.includes(struct.action)){
@@ -3992,16 +4005,17 @@ function midLoop(){
                             idx = i;
                         }
                         else {
-                            time += timeCheck(t_action,spent);
+                            time += timeCheck(t_action, global.settings.qAny ? { t: 0, r: {}} : spent);
                         }
                         global.queue.queue[i]['time'] = time;
-                        stop = true;
+                        stop = global.settings.qAny ? false : true;
                     }
                     else {
                         global.queue.queue[i].cna = true;
                         global.queue.queue[i]['time'] = -1;
                     }
                 }
+                global.queue.queue[i].qa = global.settings.qAny ? true : false;
             }
             if (idx >= 0 && c_action){
                 if (c_action.action()){
@@ -4044,6 +4058,7 @@ function midLoop(){
             for (let i=0; i<global.r_queue.queue.length; i++){
                 let struct = global.r_queue.queue[i];
                 let t_action = actions[struct.action][struct.type];
+                time = global.settings.qAny ? 0 : time;
 
                 if (t_action['grant'] && global.tech[t_action.grant[0]] && global.tech[t_action.grant[0]] >= t_action.grant[1]){
                     global.r_queue.queue.splice(i,1);
@@ -4057,16 +4072,17 @@ function midLoop(){
                             idx = i;
                         }
                         else {
-                            time += timeCheck(t_action,spent);
+                            time += timeCheck(t_action, global.settings.qAny ? { t: 0, r: {}} : spent);
                         }
                         global.r_queue.queue[i]['time'] = time;
-                        stop = true;
+                        stop = global.settings.qAny ? false : true;
                     }
                     else {
                         global.r_queue.queue[i].cna = true;
                         global.r_queue.queue[i]['time'] = -1;
                     }
                 }
+                global.r_queue.queue[i].qa = global.settings.qAny ? true : false;
             }
             if (idx >= 0 && c_action){
                 if (c_action.action()){